requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-groq
chromadb
sentence-transformers
pypdf
python-docx
unstructured
groq
faiss-cpu
tiktoken
requests
python-dotenv
plotly
pillow
beautifulsoup4
lxml
numpy